Improve broad phase scan speed, share level data instead of copying
The current broad phase scan method loads the entire world just to
determine the limits and verify validity of files. This adds a lot
of slowdown for large maps, since all files have to be read twice.

The better way to handle loading is to parse the filename to get the
X and Z coordinates of each chunk. The data is only loaded when
necssary, which greatly speeds up drawing large maps with limits.

There is also no reason to copy the level data, so the loading
process has been modified to use a boost::shared_ptr<Level> so that
each chunk is only loaded at most once.